backend/requirements.txt
# Telegram
telethon==1.37.0

# Web framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.19

# Supabase
supabase==2.11.0

# LLM & AI
anthropic==0.42.0
openai==1.59.7
langgraph==0.2.60
langchain-core==0.3.29

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2

# Utils
python-dotenv==1.0.1
httpx==0.28.1
